敵が壁を登ってくるように変更

+++ Project/Application/Object/Enemy/EnemyStatus.h
@@ -4,9 +4,9 @@ namespace Enemy {
 
 	/// @brief Enemyの現在ステータスを管理する構造体
 	struct Status {
-		int currentHealth = 100; // 現在の体力
-		int power = 1;           // Playerに与えるダメージ量
-		float moveSpeed = 5.0f;  // 移動速度
+		float currentHealth = 100.0f; // 現在の体力
+		float power = 1.0f;           // Playerに与えるダメージ量
+		float moveSpeed = 5.0f;       // 移動速度
 	};
 
 	/// @brief Enemyの種類を表す列挙型
